Require client certificate when setting a trust certificate.
Previously, it was only requested. But if the client does not provide a
certificate, OpenSSL will silently skip client authentication and grant
the connection. Based on the current MHD API, it makes no sense to only
request client authentication as it does not provide a mechanism to let
the server change its behavior if the certificate was provided or not.